pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "GeoMatchAI"
version = "0.1.0"
description = "High-Performance Visual Place Verification System"
requires-python = "==3.13.*"
dependencies = [
    "pillow>=11.3.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "timm>=1.0.22",
    "requests>=2.32.5",
    "mapillary>=1.0.15",
    "aiohttp>=3.13.2",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["geomatchai*"]

[tool.uv.sources] # TODO: make automatic selection based on available hardware
torch = [
    { index = "pytorch-cu128" },
]
torchvision = [
    { index = "pytorch-cu128" },
]
timm = [
    { index = "pypi" },
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
default = true


[tool.isort]
profile = "black"
skip_gitignore = true
known_first_party = ["geomatchai", "tests"]

[tool.ruff]
target-version = "py312"  # Use py312 as ruff doesn't support py313 yet
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "C901",  # complexity
    "E402",  # module level import not at top (needed for __init__.py to set metadata first)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "F841", "E712"]  # Don't require annotations, allow unused vars, allow == True/False in tests
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"example_usage.py" = ["N806"]  # Allow uppercase variables in example

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--verbose",
    "--cov=geomatchai",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "timm.*",
    "mapillary.*",
    "torchvision.*",
]
ignore_missing_imports = true

[dependency-groups]
plot = [
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "seaborn>=0.13.2",
]
